+++ modules/frontend/frontend.go
@@ -18,6 +18,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
+	"github.com/weaveworks/common/httpgrpc"
 	"github.com/weaveworks/common/middleware"
 	"github.com/weaveworks/common/user"
 
@@ -35,10 +36,12 @@ func NewTripperware(cfg Config, logger log.Logger, registerer prometheus.Registe
 	}, []string{"tenant"})
 
 	return func(next http.RoundTripper) http.RoundTripper {
-		// We're constructing middleware in this statement. There are two at the moment -
-		// - the rightmost one (executed first) is ShardingWare which helps to shard queries by splitting the block ID space
-		// - the leftmost one (executed last) is Deduper which dedupe Span IDs for Zipkin support
-		rt := NewRoundTripper(next, Deduper(logger), ShardingWare(cfg.QueryShards, logger))
+		// We're constructing middleware in this statement, each middleware wraps the next one from left-to-right
+		// - the Deduper dedupes Span IDs for Zipkin support
+		// - the ShardingWare shards queries by splitting the block ID space
+		// - the RetryWare retries requests that have failed (error or http status 500)
+		rt := NewRoundTripper(next, Deduper(logger), ShardingWare(cfg.QueryShards, logger), RetryWare(cfg.MaxRetries, logger))
+
 		return queryrange.RoundTripFunc(func(r *http.Request) (*http.Response, error) {
 			start := time.Now()
 			// tracing instrumentation
@@ -104,6 +107,9 @@ func NewTripperware(cfg Config, logger log.Logger, registerer prometheus.Registe
 			if resp != nil {
 				statusCode = resp.StatusCode
 				contentLength = resp.ContentLength
+			} else if httpResp, ok := httpgrpc.HTTPResponseFromError(err); ok {
+				statusCode = int(httpResp.Code)
+				contentLength = int64(len(httpResp.Body))
 			}
 
 			level.Info(logger).Log(